+++ pydgraph/__init__.py
@@ -1,6 +1,12 @@
 # SPDX-FileCopyrightText: © 2017-2025 Istari Digital, Inc.
 # SPDX-License-Identifier: Apache-2.0
 
+# Async client
+from pydgraph.async_client import AsyncDgraphClient, async_open
+from pydgraph.async_client_stub import AsyncDgraphClientStub
+from pydgraph.async_txn import AsyncTxn
+
+# Sync client
 from pydgraph.client import DgraphClient, open  # noqa: A004
 from pydgraph.client_stub import DgraphClientStub
 from pydgraph.errors import (
@@ -28,6 +34,11 @@ from pydgraph.txn import Txn
 __all__ = [
     # errors
     "AbortedError",
+    # async client
+    "AsyncDgraphClient",
+    "AsyncDgraphClientStub",
+    "AsyncTxn",
+    "async_open",
     # proto.api_pb2
     "Check",
     "ConnectionError",