+++ tests/test_queries.py
@@ -14,8 +14,7 @@ import pytest
 
 import pydgraph
 from pydgraph import open
-
-from . import helper
+from tests import helper
 
 __author__ = "Mohit Ranka <mohitranka@gmail.com>"
 __maintainer__ = "Istari Digital, Inc. <dgraph-admin@istaridigital.com>"